Fixes issue-1: Adds autoconfiguration for CassandraUnitLifecycle
