+++ gradle.properties
@@ -10,7 +10,7 @@ ktlintVersion=9.4.0
 bintrayPluginVersion=1.8.5
 
 projectGroup=com.asarkar.spring
-projectVersion=1.0.0
+projectVersion=1.0.1
 projectDescription=Starts the Cassandra server and makes the ports available as Spring environment properties
 licenseName=Apache-2.0
 licenseUrl=http://www.apache.org/licenses/LICENSE-2.0.txt
