+++ src/main/resources/META-INF/spring.factories
@@ -1,2 +1,4 @@
 org.springframework.context.ApplicationListener=\
 com.asarkar.spring.test.CassandraUnitApplicationListener
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+com.asarkar.spring.test.CassandraUnitAutoConfiguration
